Add an index soft limit and upgrade command
This works on by having each index pid report back to the manager
periodically when it doesn't have any clients. If its generation is out
of date or if the maximum number of pids is open then it will close the
index. In theory this will prefer to close idle indices over used
indices as active clients will prevent an index from checking if it
needs to close.