Fix default filters for various queries
The relation between various queries and the default filter type was a
bit more complex than I originally thought. This just maps back to the
original combinations as expected by the spatial test definitions.

It also translates the 3D bounding box with a filter case into a 400
rather than letting it bubble out as a 500.