Implement a legacy output format
This legacy format is slightly different than geojson in that it
includes top level document properties outside the "properties"
attribute. It also elides the underscore prefix for id and omits the rev
attribute completely.

This is intended as a temporary stop gap for customers that are being
upgraded from the POC geo versions to our mainline builds.

BugzId: 44517