Record geo crash
The original method to record geo crashes does not discriminte against
which easton_index process is normally stopped or crashes for some reason.
The number of reported geo crashes from bigblue MT cluster actually is
wrong and the current fix tries to discriminate against the exit types of
easton_index process.

Specifically, hasting_index is a gen_server that traps exit flag and hence
we can catch the exit reason from terminate/2 function:
1. {'EXIT', Pid, Why} is regarded as a geo crash only when Why does not
equal to normal; otherwise, it is a normal exit
2. The remainder cases of normal, shutdown, {shutdown, term()} are regarded
as normal exit

BugzID: 55017