Corrected incorrect integer_to_list call
Corrected incorrect call to integer_to_list. The check of a correct
SRID is that it is > 0, the incorrect function call was meaning this
check wasn't even being reached due to an incorrect argument being
passed to integer_to_list()

The correct two step call:

    list_to_integer(binary_to_list(Tail)),

can be replaced with a single call to

    binary_to_integer(Tail),

Also replaced the try/catch/throw mechanism for checking that the SRID specified
in the st_indexes portion of the design document is an integer. Replaced with a
guarded case statement.

CS Map checks are sufficient to validate EPSG codes. The exceptions thrown by easton
on CSMap returning a Null value for SRID validity are being caught at the very top
level in easton and only passing cryptic errors through the http responses to the user.

Correct action is to simply check that the EPSG code specified in a design document
is a positive integer in hastings and passing it through to CSMap via easton to
handle its validity. Easton will pass it to CSMap for checking. A separate task will
be raised in FogBugz to improve the handling of the thrown error on an invalid EPSG
code.

BugzID: 57093