Implement a legacy output format for Local
This legacy format is very similar to the view format however the
rows are returned under a "features" key. This matches the format
used for Local releases 1.0.0.0 to 1.0.0.4.

It is not used by default but can be enabled by setting the
hastings/default_format config variable to "legacy_local".

This is intended as a temporary stop gap for Cloudant Local
customers until we can switch them to the mainline format in the
next proper release of Local.

BugzId: 60910