Wait for index to close before destroying it
When the database is deleted index closing is racing with index
destruction (from hastings_vacuum). There is a good chance they'd both
try to acquire the internal LevelDB advisory file lock and one of them
will fail. This started happening a lot more frequently in Erlang
versions 19.0 and higher since port communication was optimized and
made more asynchronous.

Therefore when destroying the index, first check if there is an index
process, and if there is way, make sure to close it and wait for it to
exit then call destroy. The race is not completely eliminated but it
should reduce the chances.

BugzID: 106393